+++ CHANGELOG.md
@@ -71,8 +71,11 @@ _Breaking developer changes, which may affect downstream projects or sites that 
 #### :hourglass: Performance
 #### :mortar_board: Walkthrough / Help
 #### :rocket: Presets
+* For radio fields: show "unknown" values (that don't match one of the specified radio options) as a temporary placeholder raw value option ([#12082])
+* Show tooltips with tag values (and descriptions if available) for options of `radio` and `check` fields (thanks [@bhavyaKhatri2703])
 #### :hammer: Development
 * Fix unit tests failing with nodejs v26 ([#12401], thanks [@brianstrauch])
+* Remove the unused lane parsing code, which was left over from an unfinished lane editor and was not reachable from any preset ([#12664], thanks [@wantaekchoi])
 
 [#8713]: https://github.com/openstreetmap/iD/issues/8713
 [#8724]: https://github.com/openstreetmap/iD/issues/8724
@@ -81,6 +84,7 @@ _Breaking developer changes, which may affect downstream projects or sites that 
 [#10120]: https://github.com/openstreetmap/iD/pull/10120
 [#10911]: https://github.com/openstreetmap/iD/pull/10911
 [#11896]: https://github.com/openstreetmap/iD/pull/11896
+[#12082]: https://github.com/openstreetmap/iD/issues/12082
 [#12358]: https://github.com/openstreetmap/iD/pull/12358
 [#12392]: https://github.com/openstreetmap/iD/pull/12392
 [#12401]: https://github.com/openstreetmap/iD/pull/12401
@@ -100,9 +104,11 @@ _Breaking developer changes, which may affect downstream projects or sites that 
 [#12623]: https://github.com/openstreetmap/iD/pull/12623
 [#12569]: https://github.com/openstreetmap/iD/issues/12569
 [#12652]: https://github.com/openstreetmap/iD/pull/12652
+[#12664]: https://github.com/openstreetmap/iD/pull/12664
 [@brianstrauch]: https://github.com/brianstrauch
 [@jguddas]: https://github.com/jguddas
 [@greymoth-jp]: https://github.com/greymoth-jp
+[@wantaekchoi]: https://github.com/wantaekchoi
 
 
 # 2.41.2